+++ CHANGELOG.rst
@@ -8,7 +8,9 @@ Changelog
 ~~~~~~
 
 - added sign function to the backend that returns the sign of its input (1 or -1)
-- improved readthedocs documentation
+- improved readthedocs documentation (removed bug with display of math, added new use example for edge templates)
+- added a safe guard for defining edge templates: An error is raised now when edge template input variables have the
+  same name as their source variable.
 
 0.17.3
 ~~~~~~